Benchmark human and improve performance.

Benchdiff output:

```
name             old time/op    new time/op    delta
Handler/human-8     380ns ± 3%     350ns ± 3%    -7.75%  (p=0.000 n=9+10)

name             old alloc/op   new alloc/op   delta
Handler/human-8      144B ± 0%        0B       -100.00%  (p=0.000 n=10+10)

name             old allocs/op  new allocs/op  delta
Handler/human-8      3.00 ± 0%      0.00       -100.00%  (p=0.000 n=10+10)
```